Add some perspective settings This allows one to change tilt and base aka bottom of bars.
With zero tilt and negative base, perspective distortion can be entirely eliminated and all bars
stand vertical then. With 90° tilt a funny bird's-eye view can be achieved.

Additionally the size of the x*z-field can be changed. Smaller is useful for bird's-eye view and
bigger for zero tilt with rotation fixed at 0° or 90° etc.